Create Solar System Objects Database
In this project, I organized the data of solar system objects to organize and answer questions. I decided to assign these objects with common sizing labels, like tiny, small, medium, and large and this was based on mass and volume. There are many different types of objects in our solar system, so I wanted to know what was the average volume, mass, and density of a given object in the solar system. From this table, I wanted to inquiry about how many of these objects were round planets and it generated a count of 12. Lastly, I generated a list to answer the question of which objects have a mass relative to earth > .05 and objects with surface gravity of > 1.0.